refactor : 책임에 맞게 객체 분리
책임별로 분리 리팩토링 진행 및 테스트 코드 정리
- SplitDelimiter : 구분자 검증 책임
- SplitNumber : 허용 되는 값 검증 책임
- StringAddCalculator : 입력 값 합산
