+++ src/test/java/step2/StringAddCalculatorTest.java
@@ -77,73 +77,4 @@ public class StringAddCalculatorTest {
             .isInstanceOf(RuntimeException.class);
     }
 
-    @Test
-    @DisplayName("기본 구분자로 숫자 문자열 반환 검증")
-    public void splitDelimiter_쉼표구분자() {
-        // given
-        String input = "10,20,30";
-
-        // when
-        String[] result = StringAddCalculator.splitDelimiterOfDefault(input);
-
-        // then
-        assertThat(result).contains("10", "20", "30");
-    }
-
-    @Test
-    @DisplayName("기본 구분자 숫자 또는 콜론 String[] 반환 검증")
-    public void splitDelimiter_쉼표_또는_콜론구분자() {
-        // given
-        String input = "10,20:30:40";
-
-        // when
-        String[] result = StringAddCalculator.splitDelimiterOfDefault(input);
-
-        // then
-        assertThat(result).contains("10", "20", "30", "40");
-    }
-
-    @Test
-    @DisplayName("커스텀 구분자 String[] 반환 검증")
-    public void splitDelimiter_커스텀구분자() {
-        // given
-        String input = "//;\n1;2;3";
-        Matcher matcher = StringAddCalculator.CUSTOM_DELIMITER_REGEX.matcher(input);
-        String[] result = {};
-
-        // when
-        if (matcher.find()) {
-            result = StringAddCalculator.splitDelimiterOfCustom(matcher);
-        }
-
-        // then
-        assertThat(result).contains("1", "2", "3");
-    }
-
-    @Test
-    @DisplayName("stringSplit 기본 구분자 입력 반환 검증")
-    public void stringSplit_기본_구분자() {
-        // given
-        String input = "10,20:30:40";
-
-        // when
-        String[] result = StringAddCalculator.stringSplit(input);
-
-        // then
-        assertThat(result).contains("10", "20", "30", "40");
-    }
-
-    @Test
-    @DisplayName("stringSplit 커스텀 구분자 입력 반환 검증")
-    public void stringSplit_커스텀_구분자() {
-        // given
-        String input = "//;\n1;2;3";
-
-        // when
-        String[] result = StringAddCalculator.stringSplit(input);
-
-        // then
-        assertThat(result).contains("1", "2", "3");
-    }
-
 }
